alpha_signal/pyproject.toml
[project]
name = "alpha-signal"
version = "0.1.0"
description = "Turn scientific articles into actionable trading insights using LLMs"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "flytekit (>=1.16.14,<2.0.0)",
    "gcsfs (>=2026.2.0,<2027.0.0)",
    "numpy (>=2.4.2,<3.0.0)",
    "dask (>=2026.1.2,<2027.0.0)",
    "httpx (>=0.28.0)",
    "openai (>=1.60.0)",
    "pydantic (>=2.0.0)",
    "tiktoken (>=0.7.0)",
    "anthropic (>=0.84.0,<0.85.0)",
    "google-genai (>=1.65.0,<2.0.0)",
]

[project.optional-dependencies]
local = ["torch", "transformers", "peft"]

[[project.authors]]
name = "Jack DiCintio"

[project.urls]
repository = "https://github.com/jackdicintio/alpha-signal"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [{include = "alpha_signal"}]

[tool.poetry.group.dev.dependencies]
pytest = ">=8.0"
ruff = ">=0.9"
pre-commit = ">=4.0"
jupyter = "^1.1.1"
pandas = "^3.0.1"
matplotlib = "^3.10.8"
seaborn = "^0.13.2"

[tool.pytest.ini_options]
testpaths = ["alpha_signal/tests"]
markers = [
    "integration: hits live APIs (deselect with -m 'not integration')",
]

[tool.ruff]
target-version = "py312"
line-length = 100
